Updating digger speed (changing int to double and halving speed)

+++ src/main/java/frc/team2019/subsystems/Digger.java
@@ -14,7 +14,7 @@ public class Digger extends Subsystem {
         setDefaultCommand(new TeleopDigger());
     }
 
-    public void run(int speed) {
+    public void run(double speed) {
         diggerBucket.set(speed);
     }
 }
